feat: support any ci (#37)
BREAKING CHANGE: drop support `travisLaunchers` and `teamcityLaunchers` in favor of a more generic `ciLaunchers` and detection of ci using the using `is-ci` package.
BREAKING CHANGE: drop support for the `detectBrowsers` option as it was only used in one specific scenario that would cause an error anyway. Browsers are always detected when no browsers are set via the karma config or karma runner cli.